refactor(binary): move tests into dedicated modules
